FS-5062 - Combining test_db (which now only tests db/queries/application.py) with test_application
